+++ packages/cli-core/lib/command/index.ts
@@ -2,3 +2,4 @@ export * from './ArgsCommand';
 export * from './PromptsCommand';
 export * from './command.interface';
 export * from './ConfigCommand';
+export * from './InfoCommand';