+++ tools/build-tools/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tikkhun/build
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @tikkhun/cli-core@0.0.31
+
 ## 0.0.8
 
 ### Patch Changes