+++ tools/index-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tikkhun/index-generator
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @tikkhun/cli-core@0.0.31
+
 ## 0.1.4
 
 ### Patch Changes