+++ tools/lock/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tikkhun/lock
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @tikkhun/cli-core@0.0.31
+
 ## 0.0.5
 
 ### Patch Changes