+++ tools/release-plugins/info-store/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tikkhun/release-info-store
 
+## 0.0.5
+
+### Patch Changes
+
+- @tikkhun/info@0.0.7
+
 ## 0.0.4
 
 ### Patch Changes