+++ tools/release-plugins/info/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tikkhun/info
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @tikkhun/cli-core@0.0.31
+
 ## 0.0.6
 
 ### Patch Changes